pyproject.toml
[project]
name = "hr-policy-assistance"
version = "0.1.0"
description = "Intelligent HR Policy Agent using Google ADK + FastAPI"
authors = [
    {name = "shashivendraBootlabs", email = "shashivendra.singh@bootlabstech.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.12"
dependencies = [
    "fastapi",
    "uvicorn",
    "requests",
    "python-dotenv",
    "google-adk>=1.21.0,<2.0.0",
    "litellm>=1.80.11,<2.0.0",  # Gemini support
    "langchain>=0.1.0",
    "chromadb>=0.3.25",
    "pdfplumber>=0.9.0",
    "langchain-chroma (>=1.1.0,<2.0.0)",
    "langchain-ollama (>=1.0.1,<2.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "langchain-google-genai (>=4.1.2,<5.0.0)",
    "fpdf (>=1.7.2,<2.0.0)",
    "reportlab (>=4.4.7,<5.0.0)",
    "google-genai (>=1.56.0,<2.0.0)",
]


[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
